+++ validate-lib/src/test/java/no/difi/vefa/model/message/MessageTest.java
@@ -1,31 +1,23 @@
 package no.difi.vefa.model.message;
 
-import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
 public class MessageTest {
 
-    private Message message;
-
-    @Before
-    public void setUp() throws Exception {
+    @Test
+    public void constructor_normalProcedure() {
+        Message message = new Message();
         Hint hint = new Hint();
         hint.setDescription("Hint description");
         hint.setTitle("Hint title");
-
-        message = new Message();
         message.setDescription("This is my description");
         message.setMessageType(MessageType.Fatal);
         message.setSchematronRuleId("BIRULE-0123-123");
         message.setTitle("This is my title");
         message.setValidationType(ValidationType.XSL);
         message.addHint(hint);
-    }
-
-    @Test
-    public void test() {
         assertEquals("This is my description", message.getDescription());
         assertEquals(MessageType.Fatal, message.getMessageType());
         assertEquals("BIRULE-0123-123", message.getSchematronRuleId());
