+++ validate-lib/src/test/java/no/difi/vefa/utils/configuration/ConfigurationUtilsTest.java
@@ -1,24 +1,25 @@
 package no.difi.vefa.utils.configuration;
 
-import static org.junit.Assert.*;
-
 import no.difi.vefa.utils.PropertiesUtils;
-
 import org.junit.Test;
 import org.w3c.dom.Document;
 
+import java.io.File;
+
+import static org.junit.Assert.assertEquals;
+
 public class ConfigurationUtilsTest {
 
-	@Test
-	public void testFileToXMLDOM() throws Exception {
-		String path = new java.io.File("src/test/resources/validator.properties").getCanonicalPath();
-		PropertiesUtils propFile = new PropertiesUtils();
-		propFile.main(path);
-		propFile.dataDir = new java.io.File(".").getCanonicalPath();
-		
-		ConfigurationUtils configurationUtils = new ConfigurationUtils();
-		Document configDoc = configurationUtils.fileToXMLDOM(propFile.dataDir + "/src/test/resources/config.xml", propFile);
-				
-		assertEquals(configDoc.getElementsByTagName("step").item(0).getAttributes().getNamedItem("id").getNodeValue(), "XSL");
-	}
+    @Test
+    public void testFileToXMLDOM() throws Exception {
+        String path = ClassLoader.getSystemResource("validator.properties").getPath();
+        PropertiesUtils propFile = new PropertiesUtils();
+        propFile.main(path);
+        propFile.dataDir = new File(".").getCanonicalPath();
+
+        ConfigurationUtils configurationUtils = new ConfigurationUtils();
+        Document configDoc = configurationUtils.fileToXMLDOM(ClassLoader.getSystemResource("config.xml").getPath(), propFile);
+
+        assertEquals(configDoc.getElementsByTagName("step").item(0).getAttributes().getNamedItem("id").getNodeValue(), "XSL");
+    }
 }
