+++ validate-lib/src/test/resources/UBL-T10-BiiCore.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<step id="XSL" file="/src/test/resources/UBL-T10-BiiCore.xsl" />
+<step id="XSL" file="UBL-T10-BiiCore.xsl" />
