+++ validate-lib/src/test/resources/config.xml
@@ -1,15 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
 <config xmlns:xi="http://www.w3.org/2001/XInclude">
 
-	<validate 
-		id="urn:www.cenbii.eu:profile:bii04:ver1.0#urn:www.cenbii.eu:transaction:biicoretrdm010:ver1.0:#urn:www.peppol.eu:bis:peppol4a:ver1.0#urn:www.difi.no:ehf:faktura:ver1"
-		version="1.4">
-		<name>
-			<en>EHF invoice in Norway, profile invoice only</en>
-			<no>EHF faktura i Norge, profil kun faktura</no>
-		</name>		
-		<xi:include href="/src/test/resources/UBL-T10-BiiCore.xml" />		
-	</validate>	
+    <validate
+            id="urn:www.cenbii.eu:profile:bii04:ver1.0#urn:www.cenbii.eu:transaction:biicoretrdm010:ver1.0:#urn:www.peppol.eu:bis:peppol4a:ver1.0#urn:www.difi.no:ehf:faktura:ver1"
+            version="1.4">
+        <name>
+            <en>EHF invoice in Norway, profile invoice only</en>
+            <no>EHF faktura i Norge, profil kun faktura</no>
+        </name>
+        <xi:include href="/validate-lib/src/test/resources/UBL-T10-BiiCore.xml"/>
+    </validate>
 
 </config>
